Conflicts:
	.gitignore
	configure.py

+++ src/canon_perftest.cc
@@ -1,3 +1,4 @@
+#include <string.h> // strlen, strcpy
 #include "util.h"
 
 const char kPath[] =
@@ -24,9 +25,9 @@ int main() {
 
   int min = times[0];
   int max = times[0];
-  float total = 0;
+  float total = 0.0;
   for (size_t i = 0; i < times.size(); ++i) {
-    total += times[i];
+    total += static_cast<float>(times[i]);
     if (times[i] < min)
       min = times[i];
     else if (times[i] > max)
@@ -34,5 +35,5 @@ int main() {
   }
 
   printf("min %dms  max %dms  avg %.1fms\n",
-         min, max, total / times.size());
+         min, max, total / static_cast<float>(times.size()));
 }
