+++ src/parser_perftest.cc
@@ -48,7 +48,7 @@ int main(int argc, char* argv[]) {
 
       if (end - start > 100) {
         int delta = (int)(end - start);
-        float time = delta*1000 / (float)limit;
+        float time = static_cast<float>(delta) * 1000 / static_cast<float>(limit);
         printf("%s: %.1fus\n", filename, time);
         times.push_back(time);
         break;
@@ -69,7 +69,7 @@ int main(int argc, char* argv[]) {
     }
 
     printf("min %.1fus  max %.1fus  avg %.1fus\n",
-           min, max, total / times.size());
+           min, max, total / static_cast<float>(times.size()));
   }
 
   return 0;
